+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [3.1.1-alpha.3](https://github.com/zama-ai/sdk/compare/v3.1.1-alpha.2...v3.1.1-alpha.3) (2026-06-23)
+
+### Code Refactoring
+
+- **examples:** upgrade node-ethers & node-viem to @zama-fhe/sdk 3.1.0 ([#456](https://github.com/zama-ai/sdk/issues/456)) ([4cb6bc2]()), closes [#451]() [#447]()
+
+## [3.1.1-alpha.2](https://github.com/zama-ai/sdk/compare/v3.1.1-alpha.1...v3.1.1-alpha.2) (2026-06-23)
+
+### Code Refactoring
+
+- **examples:** upgrade example apps to @zama-fhe/sdk 3.1.0 ([#451](https://github.com/zama-ai/sdk/issues/451)) ([f9d233f]()), closes [#447]()
+
 ## [3.1.1-alpha.1](https://github.com/zama-ai/sdk/compare/v3.1.0...v3.1.1-alpha.1) (2026-06-22)
 
 ### Bug Fixes